requirements.txt
PyGithub>=1.59.0
requests>=2.31.0
